refactor(tui): move screen apply handlers out of Jobs
on_preview_diff, on_download_selected, on_revision_diff, on_upload_preview,
on_compact_analyze, on_restore_revision_ready, on_comments_*_loaded,
on_revisions_fetched, and on_preview_content now live as free functions
beside the screen state they mutate. Tests move with them.

Part of #383.
