+++ src/tui/bg.rs
@@ -362,8 +362,15 @@ pub(super) fn spawn_pin_push(
         state,
         "Loading diff…",
         file,
-        move |result, file, jobs, state| {
-            jobs.on_upload_preview(state, result, file, local_path, local_label, gist_label)
+        move |result, file, _jobs, state| {
+            screens::confirm::on_upload_preview(
+                state,
+                result,
+                file,
+                local_path,
+                local_label,
+                gist_label,
+            )
         },
     );
 }
@@ -386,8 +393,15 @@ pub(super) fn spawn_pin_pull(
         state,
         "Downloading…",
         file,
-        move |result, file, jobs, state| {
-            jobs.on_download_selected(state, result, target, local_label, gist_label, file)
+        move |result, file, _jobs, state| {
+            screens::diff::on_download_selected(
+                state,
+                result,
+                target,
+                local_label,
+                gist_label,
+                file,
+            )
         },
     );
 }
@@ -411,10 +425,10 @@ pub(super) fn spawn_pin_diff(
         state,
         "Loading diff…",
         file,
-        move |result, _file, jobs, state| {
+        move |result, _file, _jobs, state| {
             // Pin diffs originate from the Pins screen (no focused pane); keep the
             // historical download orientation (old = local, new = gist).
-            jobs.on_preview_diff(
+            screens::diff::on_preview_diff(
                 state,
                 result,
                 Some(local_abs),
@@ -836,13 +850,13 @@ pub(super) fn download(state: &mut AppState, mode: crate::actions::DownloadMode)
     }
 }
 
-enum LocalScanMode {
+pub(super) enum LocalScanMode {
     Flat,
     Active,
 }
 
 /// Quick re-scan used after a download/upload to make the target visible immediately.
-fn refresh_locals(
+pub(super) fn refresh_locals(
     state: &mut AppState,
     scan_mode: LocalScanMode,
     selection_target: Option<&std::path::Path>,
@@ -1224,7 +1238,9 @@ fn request_revisions(jobs: &mut Jobs, state: &mut AppState, gist_id: String) {
                 });
             (result, gist_id)
         },
-        move |(result, gist_id), jobs, state| jobs.on_revisions_fetched(state, gist_id, result),
+        move |(result, gist_id), _jobs, state| {
+            screens::revisions::on_revisions_fetched(state, gist_id, result)
+        },
     );
 }
 
@@ -1420,190 +1436,6 @@ impl Jobs {
         }
     }
 
-    /// `PreviewDiff` outcome: build the local-vs-gist preview diff and, if the two sides are
-    /// byte-identical, opportunistically refresh the pin-sync cache with the content already
-    /// fetched (see the inline comment below for why).
-    #[allow(clippy::too_many_arguments)]
-    pub(super) fn on_preview_diff(
-        &mut self,
-        state: &mut AppState,
-        result: std::result::Result<String, String>,
-        local_path: Option<PathBuf>,
-        local_label: String,
-        gist_label: String,
-        target: PathBuf,
-        upload_orientation: bool,
-    ) -> LoopFlow {
-        match result {
-            Ok(remote) => {
-                match local_path
-                    .as_ref()
-                    .map(|p| crate::domain::read_text_file_capped(p))
-                    .transpose()
-                {
-                    Ok(local) => {
-                        let local_content = local.unwrap_or_default();
-                        let diff = preview_diff_text(
-                            upload_orientation,
-                            &local_label,
-                            &local_content,
-                            &gist_label,
-                            &remote,
-                            state.ignore_trailing_newline,
-                        );
-                        let identical = crate::diff::content_eq(
-                            &local_content,
-                            &remote,
-                            state.ignore_trailing_newline,
-                        );
-                        state.enter_diff(diff, remote, local_path.unwrap_or_default(), target);
-                        set_diff_identical(state, identical);
-                        // A pin diff that turns out identical confirms the cached
-                        // last_seen_hash is (still) accurate — refresh it for free
-                        // using the content we already fetched, so the Pins list's
-                        // content-hash check (AppState::compute_pin_sync_status) stays
-                        // correct even if the gist changed elsewhere since the last
-                        // real sync. Hash the LOCAL content's raw bytes (not the
-                        // trailing-newline-normalized `identical` comparison), so
-                        // this matches the raw-byte hashing compute_pin_sync_status does.
-                        if identical {
-                            let pin = state.diff().and_then(|d| {
-                                Some((
-                                    d.gist_id.clone()?,
-                                    d.gist_filename.clone()?,
-                                    d.local_path.clone(),
-                                ))
-                            });
-                            if let Some((gid, fname, local_abs)) = pin {
-                                record_pin_sync(
-                                    state,
-                                    &local_abs,
-                                    &gid,
-                                    &fname,
-                                    &local_content,
-                                    None,
-                                );
-                            }
-                        }
-                    }
-                    Err(error) => state.set_status(format!("read failed: {error}")),
-                }
-            }
-            Err(error) => state.set_status(format!("fetch failed: {error}")),
-        }
-
-        LoopFlow::Proceed
-    }
-
-    /// `DownloadSelected` outcome: diff against an existing local file, or write a new one.
-    pub(super) fn on_download_selected(
-        &mut self,
-        state: &mut AppState,
-        result: std::result::Result<String, String>,
-        target: PathBuf,
-        local_label: String,
-        gist_label: String,
-        file: crate::domain::GistFileRef,
-    ) -> LoopFlow {
-        match result {
-            Ok(remote) => {
-                if target.exists() {
-                    match crate::domain::read_text_file_capped(&target) {
-                        Ok(local_content) => {
-                            let diff = crate::diff::unified_diff(
-                                &local_label,
-                                &local_content,
-                                &gist_label,
-                                &remote,
-                                state.ignore_trailing_newline,
-                            );
-                            let identical = crate::diff::content_eq(
-                                &local_content,
-                                &remote,
-                                state.ignore_trailing_newline,
-                            );
-                            state.staged_diff_gist =
-                                Some((file.gist_id.clone(), file.filename.clone()));
-                            state.enter_diff(diff, remote, target.clone(), target);
-                            set_diff_identical(state, identical);
-                        }
-                        Err(error) => state.set_status(error),
-                    }
-                } else {
-                    match crate::actions::execute_download(
-                        &target,
-                        &remote,
-                        crate::actions::DownloadMode::CreateNew,
-                    ) {
-                        Ok(()) => {
-                            state.set_status(format!(
-                                "Downloaded {}",
-                                target
-                                    .file_name()
-                                    .unwrap_or(target.as_os_str())
-                                    .to_string_lossy()
-                            ));
-                            record_pin_sync(
-                                state,
-                                &target,
-                                &file.gist_id,
-                                &file.filename,
-                                &remote,
-                                Some(crate::domain::SyncDirection::Download),
-                            );
-                            refresh_locals(state, LocalScanMode::Active, Some(&target));
-                        }
-                        Err(error) => state.set_status(format!("download failed: {error}")),
-                    }
-                }
-            }
-            Err(error) => state.set_status(format!("fetch failed: {error}")),
-        }
-
-        LoopFlow::Proceed
-    }
-
-    /// `UploadPreview` outcome: stage the pending Upload action and open Confirm with the
-    /// local-vs-gist diff.
-    pub(super) fn on_upload_preview(
-        &mut self,
-        state: &mut AppState,
-        result: std::result::Result<String, String>,
-        file: crate::domain::GistFileRef,
-        local_path: PathBuf,
-        local_label: String,
-        gist_label: String,
-    ) -> LoopFlow {
-        match result {
-            Ok(remote) => {
-                // Keep staged pin/list return; enter_confirm consumes it.
-                let action = PendingAction::Upload {
-                    gist_id: file.gist_id,
-                    filename: file.filename,
-                    local_path: local_path.clone(),
-                };
-                match state.init_upload_state(&local_path, Some(remote), local_label, gist_label) {
-                    Ok(()) => {
-                        // init_upload_state writes via update_upload_diff only when
-                        // Confirm is already open; open Confirm first with empty
-                        // body, then rebuild the upload diff into the payload.
-                        state.enter_confirm(action, String::new());
-                        state.update_upload_diff();
-                    }
-                    Err(error) => {
-                        state.set_status(format!(
-                            "cannot read {}: {error}",
-                            crate::config::display_path(&local_path)
-                        ));
-                    }
-                }
-            }
-            Err(error) => state.set_status(format!("fetch failed: {error}")),
-        }
-
-        LoopFlow::Proceed
-    }
-
     /// `UploadReplace` outcome: commit the pin-sync record and return to wherever the upload
     /// was initiated from, then re-fetch the gist list.
     pub(super) fn on_upload_replace(
@@ -1675,28 +1507,6 @@ impl Jobs {
         LoopFlow::Proceed
     }
 
-    /// `PreviewContent` outcome: cache the fetched content and open the read-only preview.
-    pub(super) fn on_preview_content(
-        &mut self,
-        state: &mut AppState,
-        result: std::result::Result<String, String>,
-        file: crate::domain::GistFileRef,
-        preview_title: String,
-    ) -> LoopFlow {
-        match result {
-            Ok(content) => {
-                let key = file.cache_key();
-                state
-                    .gist_content_cache
-                    .insert(key.clone(), content.clone());
-                state.enter_preview(preview_title, content, Some(key));
-            }
-            Err(error) => state.set_status(format!("fetch failed: {error}")),
-        }
-
-        LoopFlow::Proceed
-    }
-
     /// `DeleteGist` outcome.
     pub(super) fn on_delete_gist(
         &mut self,
@@ -1755,40 +1565,6 @@ impl Jobs {
         LoopFlow::Proceed
     }
 
-    /// `CompactAnalyze` outcome: a single-revision gist has nothing to compact; otherwise open
-    /// the Confirm warning before compacting.
-    pub(super) fn on_compact_analyze(
-        &mut self,
-        state: &mut AppState,
-        result: std::result::Result<usize, String>,
-        gist_id: String,
-        label: String,
-    ) -> LoopFlow {
-        match result {
-            Ok(count) if count <= 1 => state.set_status(format!(
-                "\"{label}\" already has a single revision — nothing to compact"
-            )),
-            Ok(count) => {
-                // `pending_return` was staged at the 'c' keypress (keys.rs); `enter`
-                // (inside `enter_confirm`) consumes it as the Confirm cancel path.
-                state.enter_confirm(
-                    PendingAction::CompactGist {
-                        gist_id: gist_id.clone(),
-                        label: label.clone(),
-                        count,
-                    },
-                    format!(
-                        "Compact gist {gist_id} (\"{label}\").\n\nIt has {count} revisions. Compacting clones it to a temp dir, squashes the history to a single commit, and force-pushes — the {} older revisions are gone for good.",
-                        count - 1
-                    ),
-                );
-            }
-            Err(error) => state.set_status(format!("revision check failed: {error}")),
-        }
-
-        LoopFlow::Proceed
-    }
-
     /// `CompactGist` outcome.
     pub(super) fn on_compact_gist(
         &mut self,
@@ -1805,135 +1581,6 @@ impl Jobs {
             Err(error) => state.set_status(format!("compact failed: {error}")),
         }
 
-        LoopFlow::Proceed
-    }
-
-    /// `CommentsInitialLoaded` outcome.
-    pub(super) fn on_comments_initial_loaded(
-        &mut self,
-        state: &mut AppState,
-        gist_id: String,
-        result: Result<crate::tui::InitialComments, String>,
-    ) -> LoopFlow {
-        state.apply_initial_comments(&gist_id, result);
-
-        LoopFlow::Proceed
-    }
-
-    /// `CommentsOlderLoaded` outcome.
-    pub(super) fn on_comments_older_loaded(
-        &mut self,
-        state: &mut AppState,
-        gist_id: String,
-        result: Result<Vec<GistComment>, String>,
-    ) -> LoopFlow {
-        state.apply_older_comments(&gist_id, result);
-
-        LoopFlow::Proceed
-    }
-
-    /// `RevisionsFetched` outcome. Returns [`LoopFlow::SkipIteration`] if the fetch belongs to
-    /// a gist the Revisions screen has since navigated away from.
-    pub(super) fn on_revisions_fetched(
-        &mut self,
-        state: &mut AppState,
-        gist_id: String,
-        result: std::result::Result<Vec<crate::domain::GistRevision>, String>,
-    ) -> LoopFlow {
-        if state.revision().and_then(|r| r.gist_id.as_deref()) != Some(gist_id.as_str()) {
-            return LoopFlow::SkipIteration;
-        }
-        match result {
-            Ok(entries) => {
-                let short = entries.len() <= 1;
-                if let Some(rev) = state.revision_mut() {
-                    rev.fetch_error = None;
-                    rev.entries = Some(entries);
-                }
-                if short {
-                    state.set_status("only one revision — nothing to restore");
-                }
-            }
-            Err(error) => {
-                if let Some(rev) = state.revision_mut() {
-                    rev.entries = Some(Vec::new());
-                    rev.fetch_error = Some(error);
-                }
-            }
-        }
-        LoopFlow::Proceed
-    }
-
-    /// `RevisionDiff` outcome: diff two historical revisions of the same file.
-    pub(super) fn on_revision_diff(
-        &mut self,
-        state: &mut AppState,
-        result: std::result::Result<(String, String), String>,
-        old_label: String,
-        new_label: String,
-    ) -> LoopFlow {
-        match result {
-            Ok((old_content, new_content)) => {
-                let diff = crate::diff::unified_diff(
-                    &old_label,
-                    &old_content,
-                    &new_label,
-                    &new_content,
-                    state.ignore_trailing_newline,
-                );
-                let identical = old_content == new_content;
-                // `enter_diff` (via `enter`) parks the live Revisions screen so Esc
-                // restores list cursor/entries.
-                state.enter_diff(diff, String::new(), PathBuf::new(), PathBuf::new());
-                set_diff_identical(state, identical);
-            }
-            Err(error) => state.set_status(error),
-        }
-
-        LoopFlow::Proceed
-    }
-
-    /// `RestoreRevisionReady` outcome. Returns [`LoopFlow::SkipIteration`] when the revision
-    /// content matches current (nothing to restore).
-    pub(super) fn on_restore_revision_ready(
-        &mut self,
-        state: &mut AppState,
-        result: std::result::Result<(String, String), String>,
-        gist_id: String,
-        filename: String,
-        version: String,
-        version_label: String,
-    ) -> LoopFlow {
-        match result {
-            Ok((revision_content, current_content)) => {
-                if revision_content == current_content {
-                    state.set_status("revision matches current — nothing to restore");
-                    return LoopFlow::SkipIteration;
-                }
-                let old_label = format!("revision {version_label}");
-                let new_label = format!("current {filename}");
-                let diff = crate::diff::unified_diff(
-                    &old_label,
-                    &revision_content,
-                    &new_label,
-                    &current_content,
-                    state.ignore_trailing_newline,
-                );
-                // `enter_confirm` (via `enter`) parks the live Revisions screen so
-                // cancel restores cursor/entries.
-                state.enter_confirm(
-                    PendingAction::RestoreRevision {
-                        gist_id,
-                        filename,
-                        version,
-                        version_label,
-                        content: revision_content,
-                    },
-                    diff,
-                );
-            }
-            Err(error) => state.set_status(error),
-        }
         LoopFlow::Proceed
     }
 
@@ -2019,15 +1666,6 @@ impl Jobs {
         }
 
         LoopFlow::Proceed
-    }
-}
-
-/// Set `state`'s diff-payload `identical` flag, if a diff payload is currently open. Shared by
-/// the three action outcomes that build a diff and check content identity byte-for-byte:
-/// [`Jobs::on_preview_diff`], [`Jobs::on_download_selected`], [`Jobs::on_revision_diff`].
-fn set_diff_identical(state: &mut AppState, identical: bool) {
-    if let Some(d) = state.diff_mut() {
-        d.identical = identical;
     }
 }
 
@@ -2399,47 +2037,7 @@ mod tests {
         assert!(state.status.is_none());
     }
 
-    // ---- on_preview_diff -----------------------------------------------------
-
-    #[test]
-    fn on_preview_diff_err_sets_status() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_preview_diff(
-            &mut state,
-            Err("boom".into()),
-            None,
-            "local".into(),
-            "gist".into(),
-            PathBuf::from("target"),
-            false,
-        );
-
-        assert_eq!(state.status.as_deref(), Some("fetch failed: boom"));
-    }
-
-    #[test]
-    fn on_preview_diff_ok_without_local_enters_diff() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_preview_diff(
-            &mut state,
-            Ok("remote body".into()),
-            None,
-            "local".into(),
-            "gist".into(),
-            PathBuf::from("target"),
-            false,
-        );
-
-        let diff = state.diff().expect("expected Screen::Diff");
-        assert_eq!(diff.remote_content, "remote body");
-        assert!(!diff.identical);
-    }
-
-    // ---- on_download_selected -------------------------------------------------
+    // ---- refresh_locals -------------------------------------------------
 
     #[test]
     fn refresh_locals_preserves_nested_selection_in_recursive_mode() {
@@ -2462,92 +2060,6 @@ mod tests {
         refresh_locals(&mut state, LocalScanMode::Active, Some(&target));
 
         assert_eq!(state.selected_local().map(|file| file.path), Some(target));
-    }
-
-    #[test]
-    fn on_download_selected_err_sets_status() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_download_selected(
-            &mut state,
-            Err("boom".into()),
-            PathBuf::from("target"),
-            "local".into(),
-            "gist".into(),
-            gist_file_ref("g1", "a.txt"),
-        );
-
-        assert_eq!(state.status.as_deref(), Some("fetch failed: boom"));
-    }
-
-    #[test]
-    fn on_download_selected_ok_existing_target_enters_diff() {
-        let dir = tempfile::tempdir().unwrap();
-        let target = dir.path().join("a.txt");
-        std::fs::write(&target, "local body").unwrap();
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_download_selected(
-            &mut state,
-            Ok("remote body".into()),
-            target.clone(),
-            "local".into(),
-            "gist".into(),
-            gist_file_ref("g1", "a.txt"),
-        );
-
-        let diff = state.diff().expect("expected Screen::Diff");
-        assert_eq!(diff.remote_content, "remote body");
-        // `enter_diff` takes `staged_diff_gist` and moves it onto the `DiffState` itself.
-        assert_eq!(diff.gist_id.as_deref(), Some("g1"));
-        assert_eq!(diff.gist_filename.as_deref(), Some("a.txt"));
-        assert!(state.staged_diff_gist.is_none());
-    }
-
-    // ---- on_upload_preview ------------------------------------------------------
-
-    #[test]
-    fn on_upload_preview_err_sets_status() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_upload_preview(
-            &mut state,
-            Err("boom".into()),
-            gist_file_ref("g1", "a.txt"),
-            PathBuf::from("a.txt"),
-            "local".into(),
-            "gist".into(),
-        );
-
-        assert_eq!(state.status.as_deref(), Some("fetch failed: boom"));
-    }
-
-    #[test]
-    fn on_upload_preview_ok_enters_confirm() {
-        let dir = tempfile::tempdir().unwrap();
-        let local_path = dir.path().join("a.txt");
-        std::fs::write(&local_path, "local body").unwrap();
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_upload_preview(
-            &mut state,
-            Ok("remote body".into()),
-            gist_file_ref("g1", "a.txt"),
-            local_path.clone(),
-            "local".into(),
-            "gist".into(),
-        );
-
-        assert!(state.screen.is_confirm());
-        assert!(matches!(
-            state.pending_action(),
-            Some(PendingAction::Upload { gist_id, filename, .. })
-                if gist_id == "g1" && filename == "a.txt"
-        ));
     }
 
     // ---- on_upload_replace / on_create_gist / on_delete_gist / on_remove_file /
@@ -2652,275 +2164,6 @@ mod tests {
         jobs.on_restore_revision_done(&mut state, Err("boom".into()), "g1".into(), "a.txt".into());
 
         assert_eq!(state.status.as_deref(), Some("restore failed: boom"));
-    }
-
-    // ---- on_preview_content (fully safe, both arms) ---------------------------
-
-    #[test]
-    fn on_preview_content_ok_caches_and_enters_preview() {
-        let mut state = initial_state();
-        let file = gist_file_ref("g1", "a.txt");
-        let mut jobs = empty_jobs();
-
-        jobs.on_preview_content(&mut state, Ok("body".into()), file.clone(), "a.txt".into());
-
-        assert_eq!(
-            state.gist_content_cache.get(&file.cache_key()),
-            Some(&"body".to_string())
-        );
-        let preview = state.preview().expect("expected Screen::Preview");
-        assert_eq!(preview.text, "body");
-    }
-
-    #[test]
-    fn on_preview_content_err_sets_status() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_preview_content(
-            &mut state,
-            Err("boom".into()),
-            gist_file_ref("g1", "a.txt"),
-            "a.txt".into(),
-        );
-
-        assert_eq!(state.status.as_deref(), Some("fetch failed: boom"));
-    }
-
-    // ---- on_compact_analyze (fully safe) ---------------------------------------
-
-    #[test]
-    fn on_compact_analyze_single_revision_sets_status() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_compact_analyze(&mut state, Ok(1), "g1".into(), "demo".into());
-
-        assert_eq!(
-            state.status.as_deref(),
-            Some("\"demo\" already has a single revision — nothing to compact")
-        );
-    }
-
-    #[test]
-    fn on_compact_analyze_multi_revision_enters_confirm() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_compact_analyze(&mut state, Ok(4), "g1".into(), "demo".into());
-
-        assert!(matches!(
-            state.pending_action(),
-            Some(PendingAction::CompactGist { gist_id, count, .. })
-                if gist_id == "g1" && *count == 4
-        ));
-    }
-
-    #[test]
-    fn on_compact_analyze_err_sets_status() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_compact_analyze(&mut state, Err("boom".into()), "g1".into(), "demo".into());
-
-        assert_eq!(state.status.as_deref(), Some("revision check failed: boom"));
-    }
-
-    // ---- on_comments_initial_loaded / on_comments_older_loaded (fully safe) ---
-
-    #[test]
-    fn on_comments_initial_loaded_ok_dispatches_to_detail() {
-        let mut state = initial_state();
-        state.screen = Screen::GistDetail(Box::new(DetailState {
-            gist_id: Some("g1".into()),
-            comments_loading: true,
-            ..Default::default()
-        }));
-        let mut jobs = empty_jobs();
-
-        jobs.on_comments_initial_loaded(
-            &mut state,
-            "g1".into(),
-            Ok(crate::tui::InitialComments {
-                comments: vec![GistComment {
-                    author: "alice".into(),
-                    created_at: "2026-01-01T00:00:00Z".into(),
-                    body: "hi".into(),
-                }],
-                total: 1,
-                oldest_page: 1,
-            }),
-        );
-
-        let detail = state.detail().expect("expected Screen::GistDetail");
-        assert!(!detail.comments_loading);
-        assert_eq!(detail.comments_total, Some(1));
-    }
-
-    #[test]
-    fn on_comments_older_loaded_err_dispatches_to_detail() {
-        let mut state = initial_state();
-        state.screen = Screen::GistDetail(Box::new(DetailState {
-            gist_id: Some("g1".into()),
-            comments_loading_more: true,
-            ..Default::default()
-        }));
-        let mut jobs = empty_jobs();
-
-        jobs.on_comments_older_loaded(&mut state, "g1".into(), Err("boom".into()));
-
-        let detail = state.detail().expect("expected Screen::GistDetail");
-        assert!(!detail.comments_loading_more);
-    }
-
-    // ---- on_revisions_fetched --------------------------------------------------
-
-    #[test]
-    fn on_revisions_fetched_returns_skip_iteration_on_gist_mismatch() {
-        let mut state = initial_state();
-        state.screen = Screen::Revisions(Box::new(RevisionState {
-            gist_id: Some("g1".into()),
-            ..Default::default()
-        }));
-        let mut jobs = empty_jobs();
-
-        let flow = jobs.on_revisions_fetched(&mut state, "other-gist".into(), Ok(Vec::new()));
-
-        assert!(matches!(flow, LoopFlow::SkipIteration));
-        // Not applied — still no entries.
-        assert!(state.revision().unwrap().entries.is_none());
-    }
-
-    #[test]
-    fn on_revisions_fetched_ok_sets_entries() {
-        let mut state = initial_state();
-        state.screen = Screen::Revisions(Box::new(RevisionState {
-            gist_id: Some("g1".into()),
-            ..Default::default()
-        }));
-        let entry = GistRevision {
-            version: "abc123".into(),
-            committed_at: "2026-01-01T00:00:00Z".into(),
-            user: "alice".into(),
-            change_status: crate::domain::GistRevisionChangeStatus {
-                total: 1,
-                additions: 1,
-                deletions: 0,
-            },
-        };
-        let mut jobs = empty_jobs();
-
-        let flow =
-            jobs.on_revisions_fetched(&mut state, "g1".into(), Ok(vec![entry.clone(), entry]));
-
-        assert!(matches!(flow, LoopFlow::Proceed));
-        assert_eq!(state.revision().unwrap().entries.as_ref().unwrap().len(), 2);
-    }
-
-    #[test]
-    fn on_revisions_fetched_err_sets_fetch_error() {
-        let mut state = initial_state();
-        state.screen = Screen::Revisions(Box::new(RevisionState {
-            gist_id: Some("g1".into()),
-            ..Default::default()
-        }));
-        let mut jobs = empty_jobs();
-
-        jobs.on_revisions_fetched(&mut state, "g1".into(), Err("boom".into()));
-
-        let rev = state.revision().unwrap();
-        assert_eq!(rev.fetch_error.as_deref(), Some("boom"));
-        assert_eq!(rev.entries, Some(Vec::new()));
-    }
-
-    // ---- on_revision_diff -------------------------------------------------------
-
-    #[test]
-    fn on_revision_diff_ok_enters_diff() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_revision_diff(
-            &mut state,
-            Ok(("old body".into(), "new body".into())),
-            "old".into(),
-            "new".into(),
-        );
-
-        let diff = state.diff().expect("expected Screen::Diff");
-        assert!(!diff.identical);
-    }
-
-    #[test]
-    fn on_revision_diff_err_sets_status() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_revision_diff(&mut state, Err("boom".into()), "old".into(), "new".into());
-
-        assert_eq!(state.status.as_deref(), Some("boom"));
-    }
-
-    // ---- on_restore_revision_ready ----------------------------------------------
-
-    #[test]
-    fn on_restore_revision_ready_returns_skip_iteration_when_identical() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        let flow = jobs.on_restore_revision_ready(
-            &mut state,
-            Ok(("same".into(), "same".into())),
-            "g1".into(),
-            "a.txt".into(),
-            "abc123".into(),
-            "abc1234".into(),
-        );
-
-        assert!(matches!(flow, LoopFlow::SkipIteration));
-        assert_eq!(
-            state.status.as_deref(),
-            Some("revision matches current — nothing to restore")
-        );
-    }
-
-    #[test]
-    fn on_restore_revision_ready_ok_enters_confirm_when_different() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        let flow = jobs.on_restore_revision_ready(
-            &mut state,
-            Ok(("old".into(), "new".into())),
-            "g1".into(),
-            "a.txt".into(),
-            "abc123".into(),
-            "abc1234".into(),
-        );
-
-        assert!(matches!(flow, LoopFlow::Proceed));
-        assert!(matches!(
-            state.pending_action(),
-            Some(PendingAction::RestoreRevision { gist_id, filename, .. })
-                if gist_id == "g1" && filename == "a.txt"
-        ));
-    }
-
-    #[test]
-    fn on_restore_revision_ready_err_sets_status() {
-        let mut state = initial_state();
-        let mut jobs = empty_jobs();
-
-        jobs.on_restore_revision_ready(
-            &mut state,
-            Err("boom".into()),
-            "g1".into(),
-            "a.txt".into(),
-            "abc123".into(),
-            "abc1234".into(),
-        );
-
-        assert_eq!(state.status.as_deref(), Some("boom"));
     }
 
     #[test]
